backend/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.2
Flask==2.2.2
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.0
packaging==22.0
Pillow==9.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
tokenizers==0.13.2
torch==1.9.0
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
opencv-python

